requirements.txt
huggingface_hub
huggingface_hub[hf_transfer]
hf_transfer
python-dotenv